Reorganized files into MVC pattern
Organizing the files into MVC packages will allow for easier
identification of class responsibilities.
