+++ routeFinder/model/.gitignore
@@ -1,0 +1,7 @@
+/Map$Bounds.class
+/Map$MapLoader.class
+/Map.class
+/Point.class
+/Street.class
+/AStarSearch.class
+/SearchPoint.class
