+++ routeFinder/view/.gitignore
@@ -1,0 +1,2 @@
+/MainApplet.class
+/MapView.class
